+++ client/src/components/Categories/AddExpenseField.js
@@ -2,13 +2,7 @@ import React, { useState } from "react"
 import createExpense from "../../services/createExpense"
 import { useExpenses } from "../../contexts/Expenses"
 
-interface IProps {
-	renderField: boolean,
-	setRenderField: (render: boolean) => void,
-	cId: number 
-}
-
-export default function AddExpenseField({ renderField, setRenderField, cId }: IProps) {
+export default function AddExpenseField({ renderField, setRenderField, cId }) {
 	const [expenses, setExpenses] = useExpenses()
 	const [name, setName] = useState("")
 	const [max, setMax] = useState("")
