+++ client/src/components/Categories/Category.js
@@ -3,14 +3,8 @@ import { useExpenses } from "../../contexts/Expenses"
 import CategoryHeader from './CategoryHeader'
 import AddExpenseField from './AddExpenseField'
 import Expense from "./Expense"
-import { ICategory } from '../../types/cateogry'
-import { IExpense } from '../../types/expense'
 
-interface IProps {
-	category: ICategory	
-}
-
-export default function Category({ category }: IProps) {
+export default function Category({ category }) {
 	const [expenses, ] = useExpenses()
 	const [renderAddExpenseField, setRenderAddExpenseField] = useState(false)
 
@@ -18,12 +12,14 @@ export default function Category({ category }: IProps) {
 		<React.Fragment>
             <CategoryHeader category={category} setRenderAddExpenseField={setRenderAddExpenseField} />
 			<AddExpenseField renderField={renderAddExpenseField} setRenderField={setRenderAddExpenseField} cId={category.id} />
-			{
-				expenses
-					.filter((e: IExpense) => e.categoryId === category.id)
-					.map((e: IExpense)=> <Expense key={e.id} expense={e} />)	
-			}
+			<RenderExpenses cId={category.id} expenses={expenses} />
 		</React.Fragment>
 	)
 
+}
+
+function RenderExpenses({ cId, expenses }) {
+	return expenses
+	.filter(e => e.categoryId === cId)
+	.map(e => <Expense key={e.id} expense={e} />)	
 }
