+++ client/src/components/Categories/Expense.js
@@ -1,10 +1,4 @@
-import { IExpense } from "../../types/expense"
-
-interface IProps {
-	expense: IExpense
-}
-
-export default function Expense({ expense }: IProps) {
+export default function Expense({ expense }) {
 	return (
 		<div className="h-16 px-4 flex flex-wrap justify-between items-center border-b-[1px]">
 			<h3 className="mx-8 flex justify-center items-center">{expense.name}</h3>
