+++ client/src/routes/Categories.js
@@ -1,23 +1,23 @@
 import React, { useState } from "react"
-import Header from "../components/Shared/Header"
+import Header from "../components/Header"
 import { useCategories } from "../contexts/Categories"
-import SiteWrapper from "../components/Shared/SiteWrapper"
+import SiteWrapper from "../components/SiteWrapper"
 import AddCategoryButton from "../components/Categories/AddCategoryButton"
 import AddCategoryField from "../components/Categories/AddCategoryField"
+import Footer from "../components/Footer"
 import Category from "../components/Categories/Category"
-import Protected from "../components/Shared/Protected"
-import NavigationModal from "../components/NavigationModal"
+import Protected from "../components/Protected"
 
 export default function Categories() {
 	const [renderAddCategory, setRenderAddCategory] = useState(false)
 	return (
 		<Protected>
 			<SiteWrapper>
-				<NavigationModal />
 				<Header/>
 				<AddCategoryButton setRenderAddCategory={setRenderAddCategory} />
 				<AddCategoryField renderAddCategory={renderAddCategory} setRenderAddCategory={setRenderAddCategory} />
 				<RenderCategories/>
+				<Footer/>
 			</SiteWrapper>
 		</Protected>
 	)
