+++ client/src/routes/Index.js
@@ -1,22 +1,22 @@
 /* Tailwind only works properly for .js files not .jsx */
-import Header from "../components/Shared/Header"
+import Header from "../components/Header"
+import Footer from "../components/Footer"
 import React from "react"
 import AssignMoney from "../components/Index/AssignMoney"
 import TableHeader from "../components/Index/TableHeader"
 import CategoriesAndExpensesList from "../components/Index/CategoriesAndExpensesList"
-import SiteWrapper from "../components/Shared/SiteWrapper"
-import Protected from "../components/Shared/Protected"
-import NavigationModal from "../components/NavigationModal"
+import SiteWrapper from "../components/SiteWrapper"
+import Protected from "../components/Protected"
 
 function App() {
 	return (
 		<Protected>
 			<SiteWrapper>
-				<NavigationModal />
 				<Header/>
 				<AssignMoney />
 				<TableHeader />
 				<CategoriesAndExpensesList />
+				<Footer/>
 			</SiteWrapper>
 		</Protected>
 	)
