--- client/src/types/cateogry.ts
@@ -1,7 +1,0 @@
-export interface ICategory {
-    id: number,
-    name: string,
-    updatedAt: string,
-    createdAt: string,
-    userId: string
-}
