--- client/src/types/expense.ts
@@ -1,10 +1,0 @@
-export interface IExpense {
-    id: number,
-    name: string,
-    updatedAt: string,
-    createdAt: string,
-    max: number,
-    remaining: number,
-    userId: number,
-    categoryId: number
-}
